backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
pymupdf
python-dotenv
langchain
langchain-ollama
langchain-huggingface
langchain-chroma
